+++ .ts-for-gir.gio.rc.js
@@ -1,0 +1,7 @@
+export default {
+    environments: ['gjs', 'node'],
+    modules: ['Gio-2.0'],
+    girDirectories: ['./vala-girs/gir-1.0'],
+    ignore: [],
+    promisify: true,
+}